refactor(Upgrade): port to ts
