+++ client/src/dialogs/Upgrade/UpgradeDialog.scss
