+++ client/src/dialogs/Upgrade/index.ts
@@ -1,0 +1,1 @@
+export { default } from './UpgradeDialog'
